+++ Tools.h
@@ -10,4 +10,4 @@
 
 void readFile(std::string filename, char*& input, int& size);
 void quickSortInput(char*& input);
-void quickSort(double*& hit_Xs, double*& hit_Ys, int*& hit_IDs, int*& hit_Zs, int _beginning, int _end);
+void quickSort(float*& hit_Xs, float*& hit_Ys, int*& hit_IDs, int*& hit_Zs, int _beginning, int _end);
